Changed iTextSharp to IText7 ... some changes required in code. Test passes
