+++ Frends.Community.PdfReader.sln
@@ -1,19 +1,17 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27004.2010
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29613.14
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Frends.Community.PdfReader", "Frends.Community.PdfReader\Frends.Community.PdfReader.csproj", "{3BFAD0C4-2DAC-46F9-B299-B6F9A92632D7}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Frends.Community.PdfReader", "Frends.Community.PdfReader\Frends.Community.PdfReader.csproj", "{35C305C0-8108-4A98-BB1D-AFE5C926239E}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{C89045E3-191C-4E18-B702-6357DCC55CE1}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Frends.Community.PdfReader.Tests", "Frends.Community.PdfReader.Tests\Frends.Community.PdfReader.Tests.csproj", "{8CA92187-8E4F-4414-803B-EC899479022E}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{78F7F22E-6E20-4BCE-8362-0C558568B729}"
 	ProjectSection(SolutionItems) = preProject
-		nuspec\Frends.Community.PdfReader.nuspec = nuspec\Frends.Community.PdfReader.nuspec
-		Frends.Community.PdfReader\FrendsTaskMetadata.json = Frends.Community.PdfReader\FrendsTaskMetadata.json
-		LICENSE = LICENSE
+		.gitignore = .gitignore
 		README.md = README.md
 	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Frends.Community.PdfReader.Tests", "Frends.Community.PdfReader.Tests\Frends.Community.PdfReader.Tests.csproj", "{F1D94B8C-4E73-441D-B56D-BAC600050591}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,19 +19,19 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{3BFAD0C4-2DAC-46F9-B299-B6F9A92632D7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3BFAD0C4-2DAC-46F9-B299-B6F9A92632D7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3BFAD0C4-2DAC-46F9-B299-B6F9A92632D7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3BFAD0C4-2DAC-46F9-B299-B6F9A92632D7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F1D94B8C-4E73-441D-B56D-BAC600050591}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F1D94B8C-4E73-441D-B56D-BAC600050591}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F1D94B8C-4E73-441D-B56D-BAC600050591}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F1D94B8C-4E73-441D-B56D-BAC600050591}.Release|Any CPU.Build.0 = Release|Any CPU
+		{35C305C0-8108-4A98-BB1D-AFE5C926239E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{35C305C0-8108-4A98-BB1D-AFE5C926239E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{35C305C0-8108-4A98-BB1D-AFE5C926239E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{35C305C0-8108-4A98-BB1D-AFE5C926239E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8CA92187-8E4F-4414-803B-EC899479022E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8CA92187-8E4F-4414-803B-EC899479022E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8CA92187-8E4F-4414-803B-EC899479022E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8CA92187-8E4F-4414-803B-EC899479022E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {9B1DEB4D-E6FA-4B87-AF28-641D8CC5C856}
+		SolutionGuid = {55BC6629-85C9-48D8-8CA2-B0046AF1AF4B}
 	EndGlobalSection
 EndGlobal
