+++ Frends.Community.PdfReader/xxxpackages.config
