--- nuspec/Frends.Community.PdfReader.nuspec
@@ -1,23 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<package xmlns="http://schemas.microsoft.com/packaging/2010/07/nuspec.xsd">
-  <metadata>
-    <id>Frends.Community.PdfReader</id>
-    <version>0.0.3</version>
-    <title>Frends Community PDF Reader</title>
-    <authors>HiQ Finland</authors>
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <description>Task for reading PDF files</description>
-    <summary />
-    <dependencies>
-      <dependency id="iTextSharp" version="5.5.13" />
-    </dependencies>
-    <frameworkAssemblies>
-      <frameworkAssembly assemblyName="System" targetFramework="" />
-    </frameworkAssemblies>
-  </metadata>
-  <files>
-    <file src="..\Frends.Community.PdfReader\bin\release\Frends.Community.PdfReader.dll" target="lib\net452\Frends.Community.PdfReader.dll" />
-    <file src="..\Frends.Community.PdfReader\bin\release\Frends.Community.PdfReader.xml" target="lib\net452\Frends.Community.PdfReader.xml" />
-    <file src="..\Frends.Community.PdfReader\FrendsTaskMetadata.json" target="FrendsTaskMetadata.json"/>
-  </files>
-</package>
